backend/pyproject.toml
[project]
name = "chat-backend"
version = "0.1.0"
description = "Simple chat backend with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "google-adk>=0.1.0",
    "litellm>=1.0.0",
    "kaleido>=0.2.1",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "plotly>=5.20.0",
    "python-dotenv>=1.0.1",
    "TA-Lib>=0.4.28",
    "uvicorn>=0.32.0",
    "vectorbt>=0.26.0",
    "yfinance>=0.2.40",
]

[project.optional-dependencies]
dev = [
    "uv>=0.5.0",
]
